refactor: Move API key revocation email into PublicApiKeyService
Fire-and-forget the SMTP send so the DELETE response doesn't wait on it.
