+++ packages/cli/src/controllers/__tests__/api-keys.controller.test.ts
@@ -5,22 +5,14 @@ import { mock } from 'jest-mock-extended';
 
 import { EventService } from '@/events/event.service';
 import { PublicApiKeyService } from '@/services/public-api-key.service';
-import { UrlService } from '@/services/url.service';
-import { UserManagementMailer } from '@/user-management/email';
 
 import { ApiKeysController } from '../api-keys.controller';
 
 describe('ApiKeysController', () => {
 	const publicApiKeyService = mockInstance(PublicApiKeyService);
 	const eventService = mockInstance(EventService);
-	const mailer = mockInstance(UserManagementMailer);
-	const urlService = mockInstance(UrlService);
 
 	const controller = Container.get(ApiKeysController);
-
-	beforeEach(() => {
-		urlService.getInstanceBaseUrl.mockReturnValue('https://acme.app.n8n.cloud');
-	});
 
 	describe('createAPIKey', () => {
 		it('should create and save an API key', async () => {
@@ -122,19 +114,6 @@ describe('ApiKeysController', () => {
 	});
 
 	describe('deleteAPIKey', () => {
-		const apiKey = mock<ApiKey>({
-			id: 'key-1',
-			label: 'Test 123',
-			apiKey: 'n8n_api_xxxxxxxaaa5',
-			userId: 'owner-1',
-			user: mock<User>({
-				id: 'owner-1',
-				email: 'owner@example.com',
-				firstName: 'Maria',
-				lastName: 'Silva',
-			}),
-		});
-
 		it('emits isOwn=true when the caller deletes their own key', async () => {
 			const user = mock<User>({
 				id: '123',
@@ -146,7 +125,7 @@ describe('ApiKeysController', () => {
 
 			const req = mock<AuthenticatedRequest>({ user, params: { id: user.id } });
 
-			publicApiKeyService.deleteApiKey.mockResolvedValue({ isOwn: true, apiKey });
+			publicApiKeyService.deleteApiKey.mockResolvedValue({ isOwn: true });
 
 			await controller.deleteApiKey(req, mock(), user.id);
 
@@ -156,7 +135,6 @@ describe('ApiKeysController', () => {
 				publicApi: false,
 				isOwn: true,
 			});
-			expect(mailer.apiKeyRevoked).not.toHaveBeenCalled();
 		});
 
 		it('emits isOwn=false when an admin deletes another user’s key', async () => {
@@ -167,7 +145,7 @@ describe('ApiKeysController', () => {
 
 			const req = mock<AuthenticatedRequest>({ user: admin, params: { id: 'key-of-other' } });
 
-			publicApiKeyService.deleteApiKey.mockResolvedValue({ isOwn: false, apiKey });
+			publicApiKeyService.deleteApiKey.mockResolvedValue({ isOwn: false });
 
 			await controller.deleteApiKey(req, mock(), 'key-of-other');
 
@@ -175,51 +153,6 @@ describe('ApiKeysController', () => {
 				user: admin,
 				publicApi: false,
 				isOwn: false,
-			});
-		});
-
-		it('notifies the key owner when an admin revokes their key', async () => {
-			const admin = mock<User>({
-				id: 'admin-1',
-				firstName: 'Jan',
-				lastName: 'Ostrówka',
-				email: 'jan@acme.test',
-				role: { slug: 'global:admin' },
-			});
-
-			const req = mock<AuthenticatedRequest>({ user: admin, params: { id: 'key-of-other' } });
-
-			publicApiKeyService.deleteApiKey.mockResolvedValue({ isOwn: false, apiKey });
-
-			await controller.deleteApiKey(req, mock(), 'key-of-other');
-
-			expect(mailer.apiKeyRevoked).toHaveBeenCalledWith({
-				email: 'owner@example.com',
-				firstName: 'Maria',
-				label: 'Test 123',
-				suffix: 'aaa5',
-				revokedBy: 'Jan Ostrówka',
-				revokedAt: expect.stringMatching(/^\d{1,2} [A-Z][a-z]{2} \d{4}$/),
-				createApiKeyUrl: 'https://acme.app.n8n.cloud/settings/api',
-			});
-		});
-
-		it('does not fail the request when sending the revocation email throws', async () => {
-			const admin = mock<User>({
-				id: 'admin-1',
-				firstName: 'Jan',
-				lastName: 'Ostrówka',
-				email: 'jan@acme.test',
-				role: { slug: 'global:admin' },
-			});
-
-			const req = mock<AuthenticatedRequest>({ user: admin, params: { id: 'key-of-other' } });
-
-			publicApiKeyService.deleteApiKey.mockResolvedValue({ isOwn: false, apiKey });
-			mailer.apiKeyRevoked.mockRejectedValueOnce(new Error('smtp down'));
-
-			await expect(controller.deleteApiKey(req, mock(), 'key-of-other')).resolves.toEqual({
-				success: true,
 			});
 		});
 	});
