+++ packages/cli/src/services/__tests__/public-api-key.service.integration.test.ts
@@ -25,7 +25,13 @@ describe('PublicApiKeyService', () => {
 	beforeAll(async () => {
 		await testDb.init();
 		apiKeyRepository = Container.get(ApiKeyRepository);
-		publicApiKeyService = new PublicApiKeyService(apiKeyRepository, jwtService);
+		publicApiKeyService = new PublicApiKeyService(
+			apiKeyRepository,
+			jwtService,
+			mock(),
+			mock(),
+			mock(),
+		);
 	});
 
 	afterAll(async () => {
